Myself/test_covered_form.py
import time
import pytest

from selenium.webdriver.common.by import By
from selenium.webdriver.common.keys import Keys
from datetime import datetime, timedelta
from dateutil.relativedelta import relativedelta
from perform_test import myself_flow
from config import (first_name, first_name_space, first_name_second_user, last_name, last_name_space, last_name_second_user, zip_code, zip_code_space, zip_code_invalid, employee_id, employee_id_space, employee_id_second_user, date_of_birth, date_of_birth_failed, date_of_birth_second_eser, date_of_birth_second_eser_failed, date_of_birth_not_eligible)
from conftest import driver

def test_text_match_required_all_fields(driver):
    first_name_elem, last_name_elem, zip_code_elem, employee_id_elem, date_of_birth_elem, check_coverage_button_elem, force_eligibility_elem, override_eligibility_elem = myself_flow(driver)

    expected_title = "First, let’s confirm if you're covered"
    title_text = driver.find_element(By.XPATH, '//div[@class="sc-aXZVg sc-eBMEME fcpFyo dVszMr" and contains(text(), "First, let’s confirm")]')
    assert title_text.text == expected_title
    expected_p_text = 'Check if Spark Direct is covered by your insurance or employer.'
    p_text = driver.find_element(By.CLASS_NAME, 'sc-aXZVg.sc-dCFHLb.fcpFyo.ctYaUb')
    assert p_text.text == expected_p_text
    expected_form_title = 'Your info'
    form_title = driver.find_element(By.XPATH, '//div[@class="sc-aXZVg sc-eBMEME fcpFyo dVszMr" and contains(text(), "Your info")]')
    assert form_title.text == expected_form_title
    expected_privacy_policy_text = 'Your privacy is important to us. Read more in our Privacy Policy.'
    privacy_policy_text = driver.find_element(By.CLASS_NAME, 'sc-gEvEer.clvpoc')
    assert privacy_policy_text.text == expected_privacy_policy_text

    expected_first_name_error = 'Please enter a first name'
    expected_last_name_error = 'Please enter a last name'
    expected_zip_code_error = 'Please enter a zip code'
    expected_employee_id_error = 'Please enter an employee ID'
    expected_date_of_birth_error = 'Please enter MM / DD / YYYY'
    first_name_elem.click()
    last_name_elem.click()
    zip_code_elem.click()
    employee_id_elem.click()
    date_of_birth_elem.click()
    check_coverage_button_elem.click()
    first_name_error = driver.find_element(By.ID, 'coverageFormFirstName-helper-text')
    last_name_error = driver.find_element(By.ID, 'coverageFormLastName-helper-text')
    zip_code_error = driver.find_element(By.ID, 'coverageFormZipCode-helper-text')
    employee_id_error = driver.find_element(By.ID, 'coverageFormEmployeeId-helper-text')
    date_of_birth_error = driver.find_element(By.ID, 'coverageFormDOB-helper-text')
    assert first_name_error.text == expected_first_name_error
    assert last_name_error.text == expected_last_name_error
    assert zip_code_error.text == expected_zip_code_error
    assert employee_id_error.text == expected_employee_id_error
    assert date_of_birth_error.text == expected_date_of_birth_error
    assert check_coverage_button_elem != check_coverage_button_elem.is_enabled()

def test_required_fields(driver):
    first_name_elem, last_name_elem, zip_code_elem, employee_id_elem, date_of_birth_elem, check_coverage_button_elem, force_eligibility_elem, override_eligibility_elem = myself_flow(driver)

    # first name empty
    first_name_elem.send_keys('')
    last_name_elem.send_keys(last_name)
    zip_code_elem.send_keys(zip_code)
    employee_id_elem.send_keys(employee_id)
    date_of_birth_elem.send_keys(date_of_birth)
    expected_first_name_error = 'Please enter a first name'
    first_name_error = driver.find_element(By.ID, 'coverageFormFirstName-helper-text')
    assert first_name_error.text == expected_first_name_error
    assert check_coverage_button_elem != check_coverage_button_elem.is_enabled()

    # last name empty
    first_name_elem.send_keys(first_name)
    last_name_elem.send_keys(Keys.CONTROL + 'a')
    last_name_elem.send_keys(Keys.DELETE, Keys.TAB)
    expected_last_name_error = 'Please enter a last name'
    last_name_error = driver.find_element(By.ID, 'coverageFormLastName-helper-text')
    assert last_name_error.text == expected_last_name_error
    assert check_coverage_button_elem != check_coverage_button_elem.is_enabled()

    # zip code empty
    last_name_elem.send_keys(last_name)
    zip_code_elem.send_keys(Keys.CONTROL + 'a')
    zip_code_elem.send_keys(Keys.DELETE, Keys.TAB)
    expected_zip_code_error = 'Please enter a zip code'
    zip_code_error = driver.find_element(By.ID, 'coverageFormZipCode-helper-text')
    # Figma Zip code is separated Zip code error text is present together without space
    # Requires test fix after bug fix (zip_code_error.text == expected_zip_code_error)
    assert zip_code_error.text != expected_zip_code_error
    assert check_coverage_button_elem != check_coverage_button_elem.is_enabled()

    # employee ID empty
    zip_code_elem.send_keys(zip_code)
    employee_id_elem.send_keys(Keys.CONTROL + 'a')
    employee_id_elem.send_keys(Keys.DELETE, Keys.TAB)
    expected_employee_id_error = 'Please enter an employee ID'
    employee_id_error = driver.find_element(By.ID, 'coverageFormEmployeeId-helper-text')
    assert employee_id_error.text == expected_employee_id_error
    assert check_coverage_button_elem != check_coverage_button_elem.is_enabled()

    # date of birth empty
    employee_id_elem.send_keys(employee_id)
    date_of_birth_elem.send_keys(Keys.CONTROL + 'a')
    date_of_birth_elem.send_keys(Keys.DELETE, Keys.TAB)
    expected_date_of_birth_error = 'Please enter MM / DD / YYYY'
    date_of_birth_error = driver.find_element(By.ID, 'coverageFormDOB-helper-text')
    assert date_of_birth_error.text == expected_date_of_birth_error
    assert check_coverage_button_elem != check_coverage_button_elem.is_enabled()


def test_date_of_birth_field_validation(driver):
    first_name_elem, last_name_elem, zip_code_elem, employee_id_elem, date_of_birth_elem, check_coverage_button_elem, force_eligibility_elem, override_eligibility_elem = myself_flow(driver)

    first_name_elem.send_keys(':)')
    last_name_elem.send_keys(':)')
    zip_code_elem.send_keys('12345')
    employee_id_elem.send_keys('12345')

    # Please confirm if the error message for users under 18 should be the same for the A dependent and Myself flow or should be removed some text in the Myself flow with regards to the dependent.
    # need update Test regarding the answer
    today = datetime.now().date()
    day_under_13 = today - relativedelta(years=13, days=-1)
    formatted_day_under_13 = day_under_13.strftime('%m/%d/%Y')
    date_of_birth_elem.send_keys(formatted_day_under_13)
    date_of_birth_error_under18 = driver.find_element(By.ID, 'coverageFormDOB-helper-text')
    expected_date_of_birth_under18 = 'Spark Direct is not intended for users under the age of 13. If your dependent is under the age of 13, they are not eligible to use Spark Direct.'
    assert date_of_birth_error_under18.text == expected_date_of_birth_under18
    assert check_coverage_button_elem.is_enabled()

    # need update Test regarding the answer
    date_of_birth_elem.send_keys(Keys.CONTROL + 'a')
    invalid_date_of_birth = {
        # '13 / 12 / 2010',
        # '01.01.2000',
        # '01-01-2000',
        '13/12/2000',
        '09/31/2000',
        '11/31/2000',
        '12/32/2000',
        '02 / 29 / 2001',
        '02/29/2001',
        '02/28/1800',
        '04/31/2000',
        '04/30/1800',
    }
    for date in invalid_date_of_birth:
        if date_of_birth_elem.get_attribute("value"):
            date_of_birth_elem.send_keys(Keys.CONTROL + 'a')
        date_of_birth_elem.send_keys(date)
        check_coverage_button_elem.click()
        expected_date_of_birth = 'Please enter MM / DD / YYYY'
        date_of_birth_error = driver.find_element(By.ID, 'coverageFormDOB-helper-text')
        assert date_of_birth_error.text == expected_date_of_birth
        assert check_coverage_button_elem != check_coverage_button_elem.is_enabled()

    # need update Test regarding the answer
    today = datetime.now().date()
    years_old_13 = today - relativedelta(years=13)
    formatted_13_years_old = years_old_13.strftime('%m/%d/%Y')
    valid_date_of_birth = {
        formatted_13_years_old,
        '01/01/2000',
        '02/29/2000',
        '05/31/2000',
        '01/01/1900',
        '12/31/2009',
        '12/31/2008',
        # '01.01.2000',
        # '01-01-2000',
        # '01 / 01 / 2000',
    }
    for date_valid in valid_date_of_birth:
        if date_of_birth_elem.get_attribute('value'):
            date_of_birth_elem.send_keys(Keys.CONTROL + 'a')
        date_of_birth_elem.send_keys(date_valid)
        assert check_coverage_button_elem.is_enabled()

def test_zip_code_field_validation(driver):
    first_name_elem, last_name_elem, zip_code_elem, employee_id_elem, date_of_birth_elem, check_coverage_button_elem, force_eligibility_elem, override_eligibility_elem = myself_flow(driver)

    expected_zip_code_error = 'Please enter a zip code'
    first_name_elem.send_keys(':)')
    last_name_elem.send_keys(':)')
    employee_id_elem.send_keys('12345')
    date_of_birth_elem.send_keys('01/01/2000')

    # need update Test regarding the answer
    # error text is zipcode not zip code
    invalid_zip_code = [
        '     ',
        '_____',
        '1234r',
        '!@#$%',
        '123456',
        '12345-123',
        '12345 12345',
        '12345 pppp',
        'ppppp',
        '123 45',
        '123 4',
        '12345 '
    ]
    for data in invalid_zip_code:
        if zip_code_elem.get_attribute('value'):
            zip_code_elem.send_keys(Keys.CONTROL + 'a')
        zip_code_elem.send_keys(data)
        # check_coverage_button.click()
        # zip_code_error = driver.find_element(By.ID, 'coverageFormZipCode-helper-text')
        # assert zip_code_error.text == expected_zip_code_error
        assert check_coverage_button_elem != check_coverage_button_elem.is_enabled()

    valid_zip_code = [
        '12345',
        '11111',
        '12345-1234',
        '12345 1234',
    ]
    for data_valid in valid_zip_code:
        if zip_code_elem.get_attribute('value'):
            zip_code_elem.send_keys(Keys.CONTROL + 'a')
        zip_code_elem.send_keys(data_valid)
        assert check_coverage_button_elem.is_enabled()

def test_privacy_policy_newtab(driver):
    myself_flow(driver)

    # need to update test
    # link opens in the same tab
    privacy_policy = driver.find_element(By.CLASS_NAME, 'sc-tagGq.ivXcxM')
    privacy_policy.click()
    privacy_link_url = 'https://www.bighealth.com/spark-direct-privacy-policy/'
    # driver.switch_to.window(driver.window_handles[2])
    assert privacy_link_url == driver.current_url
